Refactor tests to share functionability
